few more updates for method implementation
